1.5.9: fix a crash on startup introduced in 1.5.1
iHideKeyCode and iZoomToggleKeyCode were registered with std::uint32_t values
while their names begin with "i". MakeSetting takes the setting's type from that
first letter - i signed, u unsigned, f float, b bool, s string - and when the
value does not match it deletes the name and leaves it null rather than failing
loudly. The game's INISettingCollection then dereferences that name while
inserting, and Skyrim dies during SKSEPluginLoad.

Three things agreed on the diagnosis: DragonsEyeMinimap.log stopped after
"Loading ..." and never reached "Succesfully loaded!"; skse64.log's final line
was "checking plugin ... DragonsEyeMinimap.dll" with no result and no further
plugins, so nothing alphabetically after D was ever reached; and the two
offending calls are visible in Settings.cpp.

Both settings are std::int32_t now, matching their prefix.

Every setting is also registered through AddChecked, which refuses one whose name
came back null and says which it was. The failure mode for this class of mistake
is now a line in the log rather than a crash with nothing to go on.

Not a 1.5.97 problem: SKSE 2.0.20 against runtime 01050610 is correct for that
version, and the DLL still resolves through versionlib and knows the SE runtime
names.

+++ source/UI.cpp
@@ -133,7 +133,7 @@ namespace UI
 				return false;
 			}
 
-			settings::controls::hideKeyCode = buttonEvent->GetIDCode();
+			settings::controls::hideKeyCode = static_cast<std::int32_t>(buttonEvent->GetIDCode());
 			awaitingKeyBind.store(false);
 
 			// Swallow it, so binding a key does not also trigger whatever it is bound to.
@@ -309,10 +309,10 @@ namespace UI
 			}
 			HelpMarker("Zoom the map where you want it, then press \"Set to current\" to store that level as a preset.");
 
-			int zoomKey = static_cast<int>(controls::zoomToggleKeyCode);
+			int zoomKey = controls::zoomToggleKeyCode;
 			if (ImGuiMCP::InputInt("Zoom toggle key", &zoomKey))
 			{
-				controls::zoomToggleKeyCode = static_cast<std::uint32_t>(zoomKey < 0 ? 0 : zoomKey);
+				controls::zoomToggleKeyCode = zoomKey < 0 ? 0 : zoomKey;
 			}
 			HelpMarker("Tapping this key jumps between the two presets, instead of holding the control key and scrolling. 0 disables it.");
 
@@ -328,10 +328,10 @@ namespace UI
 
 			ImGuiMCP::SeparatorText("Controls");
 
-			int keyCode = static_cast<int>(controls::hideKeyCode);
+			int keyCode = controls::hideKeyCode;
 			if (ImGuiMCP::InputInt("Hide key", &keyCode))
 			{
-				controls::hideKeyCode = static_cast<std::uint32_t>(keyCode < 0 ? 0 : keyCode);
+				controls::hideKeyCode = keyCode < 0 ? 0 : keyCode;
 			}
 			HelpMarker("DirectInput scan code of a key that shows or hides the minimap the moment it is pressed. "
 					   "0 disables it. This is separate from the game's Local Map key, which keeps its own "
